Various fixes for narrowing conversion warnings
Fixes for clang-tidy bugprone-narrowing-conversions check.

+++ src/time_management.h
@@ -19,7 +19,7 @@ public:
   AlphaZeroTimeManager() = default;
   AlphaZeroTimeManager(float percentage) : percentage_(percentage) {}
   float budget_ms(int time_left_ms, std::optional<int> inc_ms, const board::Board& b) {
-    return (time_left_ms - move_overhead_ms_) * percentage_ * 0.01;
+    return (static_cast<float>(time_left_ms) - move_overhead_ms_) * percentage_ * 0.01f;
   }
 };
 
@@ -36,12 +36,12 @@ public:
   float budget_ms(int time_left_ms, std::optional<int> inc_ms, const board::Board& b) {
     int increment = inc_ms ? std::max(0, *inc_ms) : 0;
 
-    float time_available = time_left_ms - move_overhead_ms_;
+    float time_available = static_cast<float>(time_left_ms) - move_overhead_ms_;
 
     float time_ratio = static_cast<float>(increment) / static_cast<float>(time_left_ms);
 
     // Increase percentage as ply count increases.
-    float frac = (base_pct_ + b.total_moves * ply_pct_) * 0.01;
+    float frac = (base_pct_ + static_cast<float>(b.total_moves) * ply_pct_) * 0.01f;
 
     // std::cout << "frac before: " << frac << "\n";
     // Increase fraction as ratio of increment to total time reaches equality
